backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Supabase client (used for Storage presigning + admin auth operations)
supabase==2.10.0

# JWT validation — Supabase issues standard HS256 JWTs
python-jose[cryptography]==3.3.0

# HTTP client for calling n8n webhooks and external APIs (pinned to <0.28 for supabase compat)
httpx==0.27.2

# Stripe SDK
stripe==11.3.0

# Environment variable management
python-dotenv==1.0.1

# Password hashing (for temporary passwords during onboarding invite)
passlib[bcrypt]==1.7.4

# Symmetric encryption for access codes at rest (Fernet = AES-128-CBC + HMAC)
cryptography==44.0.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
